refactor(flip-ui): remove dead AiSelect/AiChipSelect and @popperjs/core (#1063)
Neither component is rendered anywhere, and the codebase's only
@popperjs/core import existed to serve them. 705 lines of dead code.

The two arrived here differently, which is worth recording before the
code is gone:

  - AiChipSelect was real. It was the multi-select role picker in
    RegisterUserModal until 84a2004f ("feat(user-profile): name +
    organisation on users"), which changed the modal from multi-role to
    single-role and replaced it with a Headless UI Listbox written inline
    in the template. The removal was collateral from a redesign, not a
    deliberate deprecation, so multi-select role assignment is a
    capability the platform quietly lost; this component is the prior art
    if it is ever wanted back.
  - AiSelect was never used in this repo at all. A pickaxe for `<AiSelect`
    across all history returns nothing, and at the 0d3afae7 migration
    commit its only importer was already its own spec. It arrived dead
    from the predecessor repo.

Both carried green unit tests, which is why neither was ever flagged.

IOption outlives them — RegisterUserModal's inline Listbox still uses it —
so it moves from src/components/AiSelect/interfaces.ts to
src/interfaces/select.ts rather than being deleted. Leaving it behind
would have kept a directory named after two components that no longer
exist.

@popperjs/core leaves package.json entirely. It stays installed as a
transitive dependency of tippy.js, which is correct: nothing in our source
imports it directly any more. This also retires the judgement call it
forced in #1041, where it was the one package whose stanza could not be
settled by "does it ship" — it is now settled by "nothing imports it".

Also removes two stale references the deletion makes vacuous: the
assertion in RegisterUserModal.spec.ts that data-test="chip-select" is
absent (it can no longer be present, so the assertion could never fail),
and the aside in the Cypress manage_users spec describing the superseded
chip-select wording.

The production bundle is byte-identical to the parent branch — the dead
code was already tree-shaken out, so this is a pure source-level cleanup
with no runtime effect.

Note the ESLint dependency-scoping rule added in #1041 does not catch this
class of drift: it checks that what IS imported is declared, not that what
is declared is reachable. Dead-code sweeps stay manual.
