+++ flip-ui/package.json
@@ -39,7 +39,6 @@
     },
     "dependencies": {
         "@headlessui/vue": "^1.7.0",
-        "@popperjs/core": "^2.10.2",
         "@vueuse/core": "^8.7.5",
         "aws-amplify": "^6.15.3",
         "axios": "^1.16.0",
