--- flip-ui/src/utils/popper.ts
@@ -1,47 +1,0 @@
-/*
- * Copyright (c) 2026 Guy's and St Thomas' NHS Foundation Trust & King's College London
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
-
-
-import { createPopper } from "@popperjs/core";
-import { OptionsGeneric } from "@popperjs/core/lib/types";
-import { onMounted, Ref, ref, watchEffect } from "vue";
-
-export function usePopper(options: Partial<OptionsGeneric<Partial<unknown>>>): {
-    trigger: Ref<HTMLElement | null>;
-    popper: Ref<HTMLElement | null>
-} {
-    const trigger = ref<HTMLElement | null>(null);
-    const popper = ref<HTMLElement | null>(null);
-
-    onMounted(() => {
-        watchEffect(onInvalidate => {
-            const popperValue = popper.value as HTMLElement & { el?: HTMLElement } | null;
-            const triggerValue = trigger.value as HTMLElement & { el?: HTMLElement } | null;
-            const popperEl = popperValue?.el ?? popperValue;
-            const referenceElement = triggerValue?.el ?? triggerValue;
-            if (!(referenceElement instanceof HTMLElement) || !(popperEl instanceof HTMLElement)) {
-                return;
-            }
-
-            const { destroy } = createPopper(referenceElement, popperEl, options);
-            onInvalidate(destroy);
-        });
-    });
-
-    return {
-        trigger,
-        popper
-    };
-}
